commit ajustes na edição e criação do dic com indicadores externos associados;.

+++ src/main/java/br/gov/es/siscap/repository/IndicadorExternoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import br.gov.es.siscap.models.IndicadorExterno;
 
-public interface IndicadorExternoRepository extends JpaRepository<IndicadorExterno, Long> {
+public interface IndicadorExternoRepository extends JpaRepository<IndicadorExterno, Integer> {
 
     @Query("""
                 SELECT DISTINCT ie
